demultiplex table
