+++ server/reactor/poll_reactor_impl.cpp
@@ -34,70 +34,6 @@ short reactor::reactor_event_to_poll_event(event_handler::Event_Type type, int p
         // CONNECT_EVENT   = 1 << 7
 }
 
-int poll_event_repo::bind_new(Event_Type type, event_handler* handler)
-{
-    if(type == event_handler::NONE || handler == 0)
-    {
-        LOG(WARNING) << "can't bind NONE event or handler is nullptr";
-        return -1;
-    }
-
-    types_and_handlers_.push_back(event_pair{type, handler});
-    return 0;
-}
-
-int poll_event_repo::unbind(Event_Type type, const event_handler* handler)
-{
-    if(type == event_handler::NONE || handler == 0)
-    {
-        LOG(WARNING) << "can't unbind NONE event or handler can't be nullptr";
-        return -1;
-    }
-
-    auto event_tuple_find = this->find(type);
-    if(event_tuple_find == types_and_handlers_.end())
-    {
-        LOG(WARNING) << "can't unbind, no this type of event";
-        return -1;
-    }
-
-    if(event_tuple_find->handler != handler)
-    {
-        LOG(WARNING) << "the handler you unbinded is not the same as you provided";
-        return -1;
-    }
-
-    types_and_handlers_.erase(event_tuple_find);
-    return 0;
-}
-
-event_handler* poll_event_repo::get_handler(Event_Type type) const 
-{
-    auto iter = this->find(type);
-    if(iter == types_and_handlers_.end())
-    {
-        // LOG(WARNING) << "can't unbind, no this type of event";
-        return nullptr;
-    }
-    return iter->handler;
-}
-
-std::vector<poll_event_repo::event_tuple>::const_iterator poll_event_repo::find(Event_Type type) const
-{
-//    size_t size = types_and_handlers_.size();
-    return std::find_if(types_and_handlers_.begin(), types_and_handlers_.end(), 
-        [type] (const event_tuple& tuple) 
-        {
-            if((tuple.handle == event_handler::READ_EVENT || tuple.handle == event_handler::ACCEPT_EVENT) 
-                && type == POLLIN)
-                return true;
-            if((tuple.handle == event_handler::WRITE_EVENT || tuple.handle == event_handler::CONNECT_EVENT)
-                && type == POLLOUT)
-                return true;
-            if((tuple.handle == event_handler::EXCEPT_EVENT && type == POLLPRI)) return true;
-			return false;
-        });
-}
 
 poll_reactor_impl::poll_reactor_impl() : wait_pfds_(), demux_table_() { }
