+++ server/reactor/poll_reactor_impl.h
@@ -1,101 +1,19 @@
 #ifndef _UNP_POLL_REACTOR_IMPL_H_
 #define _UNP_POLL_REACTOR_IMPL_H_
 
-#include "server/reactor/reactor_implementation.h"
-#include "server/util/easylogging++.h"
-#include "server/reactor/select_reactor_impl.h"
-#include "boost/shared_ptr.hpp"
-#include "boost/make_shared.hpp"
 #include <poll.h>
 #include <sys/epoll.h>
 #include <chrono>
 #include <mutex>
+
+#include "server/reactor/reactor_implementation.h"
+#include "server/util/easylogging++.h"
 #include "server/reactor/epoller.h"
+#include "server/reactor/poll_demultiplex_table.h"
 
 using namespace std::chrono_literals;
 
 namespace reactor{
-
-struct event_pair
-{
-    event_pair() : handle(0), handler(nullptr){}
-    event_pair(unsigned int handle, event_handler* handler)
-        : handle(handle), handler(handler){}
-    unsigned int handle;
-    event_handler* handler;
-};
-
-class poll_event_repo {
-public:
-    // using event_tuple = std::pair<unsigned int, event_handler*>;
-    using event_tuple = event_pair;
-    using Event_Type = event_handler::Event_Type;
-    poll_event_repo() : types_and_handlers_() {}
-    ~poll_event_repo() {}
-    int bind_new(Event_Type type, event_handler* handler);
-    int unbind(Event_Type type, const event_handler* handler);
-    void clear() {types_and_handlers_.clear();}
-    event_handler* get_handler(Event_Type type) const ;
-    int handle_count() const {return types_and_handlers_.size();};
-
-private:
-    std::vector<event_tuple>::const_iterator find(Event_Type type) const;
-private:
-    std::vector<event_tuple> types_and_handlers_;
-};
-
-class poll_demultiplex_table {
-public:
-    using Event_Type = event_handler::Event_Type;
-    poll_demultiplex_table() : table_(), size_(0){}
-    event_handler* get_handler(int handle, Event_Type type) const 
-    {
-        if(table_.size() <= static_cast<size_t>(handle))
-            return nullptr;
-        return table_[handle].get_handler(type);
-    }
-
-    bool has_handle(int handle) const 
-    {
-        return table_.size() > static_cast<size_t>(handle) && table_[handle].handle_count() > 0;
-    }
-
-    const std::vector<poll_event_repo>& get_event_vector() const 
-    { 
-        return table_;
-    }
-
-    int bind(int handle, event_handler* handler, Event_Type type){
-        if(static_cast<size_t>(handle) >= table_.size())
-        {
-            table_.resize(handle + 10);
-        }
-        size_++;
-        LOG(INFO) << "size_: " << size_;
-        return table_[handle].bind_new(type, handler);
-    }
-
-    //unbind掉绑定到这个handle的所有事件处理器
-    int unbind(int handle){
-        int handle_count = table_[handle].handle_count();
-        if(handle_count <= 0)
-        {
-            LOG(WARNING) << "No handler bind to handle: " << handle;
-            return -1;
-        }
-        table_[handle].clear();
-        size_ -= handle_count;
-        LOG(INFO) << "unbinding size_: " << size_;
-        return 0;
-    }
-    int unbind(int handle, const event_handler* handler, Event_Type type){
-        return table_[handle].unbind(type, handler);
-    }
-private:
-    std::vector<poll_event_repo>    table_;
-    int                             size_;
-};
-
 
 class poll_reactor_impl : public reactor_implementation{
 public:
