+++ server/reactor/reactor_implementation.h
@@ -3,12 +3,13 @@
 
 #include "server/reactor/event_handler.h"
 #include <chrono>
-#include <atomic>
+#include <mutex>
 namespace reactor{
 
 class reactor_implementation{
 public:
     using Event_Type = event_handler::Event_Type;
+    using mutex_t = std::mutex;
     reactor_implementation() = default;
     virtual ~reactor_implementation(){}
     virtual int handle_events(std::chrono::microseconds *timeout) = 0;
@@ -18,8 +19,8 @@ public:
     virtual int unregister_handler(int handle, event_handler *handler, Event_Type type) = 0;
     bool isWaiting() const {return isWaiting_;}
 protected:
-    std::atomic_bool isWaiting_ = false;
+    bool        isWaiting_ = false;
+    mutex_t     mutex_;
 };
-
 }
 #endif //_UNP_REACTOR_IMPLEMENTATION_H_
